Move sorting directories under algorithms directory
