+++ algorithms/sorting/Insertion Sort/insertion-sort.js
@@ -1,4 +1,4 @@
-const util = require( '../../util' );
+const util = require( '../../../util' );
 
 /**
  * @description Performs an insertion sort on the provided array. Insertion sort is an stable, in-place sorting algorithm. It
